library/src/test/java/it/unicam/cs/pa/robotswarmsim/library/robotcommand/TimedCommandTest.java
package it.unicam.cs.pa.robotswarmsim.library.robotcommand;

import it.unicam.cs.pa.robotswarmsim.library.Environment;
import it.unicam.cs.pa.robotswarmsim.library.Point;
import it.unicam.cs.pa.robotswarmsim.library.entity.Robot;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

/**
 * Test class for TimedCommand.
 */
public class TimedCommandTest {

    private static final double DELTA = 0.0001; // Precision for double comparisons
    private MockTimedCommand command;
    private Robot robot;

    /**
     * Sets up the testing environment.
     * This method is called before each test. It initializes a mock TimedCommand and a mock Robot.
     */
    double x = 0.0; // An exemplary value for the x-coordinate
    double y = 0.0; // An exemplary value for the y-coordinate

    @BeforeEach
    public void setUp() {
        Point mockPoint = new Point(x, y); // Create a point with x and y coordinates
        Environment mockEnvironment = new Environment(/* parameters if needed */);
        robot = new Robot(mockPoint, mockEnvironment); // Create a Robot with parameters
        command = new MockTimedCommand(5.0); // Command with 5 seconds execution time
    }

    /**
     * Tests that the command executes correctly over time.
     */
    @Test
    public void testTimedCommandExecution() {
        double dt = 1.0; // 1 second time step

        for (int i = 0; i < 4; i++) {
            command.execute(robot, dt);
            assertEquals(i + 1, robot.getTimedCommandTime(), DELTA, "Elapsed time should be correct after execution.");
            assertFalse(command.isCompleted(robot.getTimedCommandTime()), "Command should not be completed yet.");
        }

        command.execute(robot, dt);
        assertEquals(5.0, robot.getTimedCommandTime(), DELTA, "Elapsed time should match the execution time.");
        assertTrue(command.isCompleted(robot.getTimedCommandTime()), "Command should be completed.");
    }

    /**
     * Mock implementation of TimedCommand for testing purposes.
     */
    private static class MockTimedCommand extends TimedCommand {

        public MockTimedCommand(double executionTime) {
            super(executionTime);
        }

        @Override
        protected void performCommand(Robot robot, double dt) {
            // Mock implementation of command logic
            // Add logic here to simulate command behavior, if necessary
        }
    }
}
